refactor!: change Event structure to four-momenta
Define an Event as a list of four-momenta and a weight, instead of a
list of particles and weight. The particle information of the Event is a
constant across all Events belonging to this specific reaction.
Therefore extract this info and define it in a header-like part of a
EventList.

* Change Event structure and introduce EventList structure
* Remove Particle class and rename files
* Change data and kinematics interfaces accordingly
* Remove DataCorrection functionality, as this is a special part of data
  prepartion and does not belong into ComPWA